moved EventMapper to mapper.event package
