+++ src/main/java/school/faang/user_service/service/event/EventParticipationService.java
@@ -3,7 +3,7 @@ package school.faang.user_service.service.event;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import school.faang.user_service.dto.UserDto;
-import school.faang.user_service.mapper.UserMapper;
+import school.faang.user_service.mapper.event.UserMapper;
 import school.faang.user_service.repository.event.EventParticipationRepository;
 
 import java.util.List;
